nnstreamer: add missing protobuf files generation
This patch manually invokes protoc to generate the missing model_flags.pb.h,
types.pb.h, and error_codes.pb.h headers, which are necessary for successful
compilation of nnstreamer.

While this approach resolves the immediate issue, it is a temporary
workaround and should be revisited to integrate a more maintainable solution
in the future.

Signed-off-by: Bruno Bonaldi Teixeira <bruno.t@variscite.com>